+++ gr-grpc/CMakeLists.txt
@@ -21,7 +21,7 @@
 # Project setup
 ########################################################################
 cmake_minimum_required(VERSION 3.1 FATAL_ERROR)
-project(gr-grpcall CXX C)
+project(gr-grgrpc CXX C)
 enable_testing()
 
 # Enable C++11 support
@@ -93,7 +93,7 @@ set(Boost_ADDITIONAL_VERSIONS
 find_package(Boost "1.35" COMPONENTS filesystem system)
 
 if(NOT Boost_FOUND)
-    message(FATAL_ERROR "Boost required to compile grpcall")
+    message(FATAL_ERROR "Boost required to compile grgrpc")
 endif()
 
 ########################################################################
@@ -102,7 +102,7 @@ endif()
 include(GrPlatform) #define LIB_SUFFIX
 set(GR_RUNTIME_DIR      bin)
 set(GR_LIBRARY_DIR      lib${LIB_SUFFIX})
-set(GR_INCLUDE_DIR      include/grpcall)
+set(GR_INCLUDE_DIR      include/grgrpc)
 set(GR_DATA_DIR         share)
 set(GR_PKG_DATA_DIR     ${GR_DATA_DIR}/${CMAKE_PROJECT_NAME})
 set(GR_DOC_DIR          ${GR_DATA_DIR}/doc)
@@ -149,11 +149,11 @@ list(INSERT CMAKE_MODULE_PATH 0 ${CMAKE_SOURCE_DIR}/cmake/Modules)
 include(GrVersion)
 
 if(NOT CPPUNIT_FOUND)
-    message(FATAL_ERROR "CppUnit required to compile grpcall")
+    message(FATAL_ERROR "CppUnit required to compile grgrpc")
 endif()
 
 ########################################################################
-# Find grpcall dependencies
+# Find grgrpc dependencies
 ########################################################################
 find_package(Protobuf REQUIRED)
 message(STATUS "Using protobuf ${protobuf_VERSION}")
@@ -190,8 +190,8 @@ link_directories(
 )
 
 # Set component parameters
-set(GR_GRPCALL_INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/include CACHE INTERNAL "" FORCE)
-set(GR_GRPCALL_SWIG_INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/swig CACHE INTERNAL "" FORCE)
+set(GR_GRGRPC_INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/include CACHE INTERNAL "" FORCE)
+set(GR_GRGRPC_SWIG_INCLUDE_DIRS ${CMAKE_CURRENT_SOURCE_DIR}/swig CACHE INTERNAL "" FORCE)
 
 ########################################################################
 # Create uninstall target
@@ -208,7 +208,7 @@ add_custom_target(uninstall
 ########################################################################
 # Add subdirectories
 ########################################################################
-add_subdirectory(include/grpcall)
+add_subdirectory(include/grgrpc)
 add_subdirectory(lib)
 add_subdirectory(swig)
 add_subdirectory(python)
@@ -223,6 +223,6 @@ if(NOT CMAKE_MODULES_DIR)
   set(CMAKE_MODULES_DIR lib${LIB_SUFFIX}/cmake)
 endif(NOT CMAKE_MODULES_DIR)
 
-install(FILES cmake/Modules/grpcallConfig.cmake
-    DESTINATION ${CMAKE_MODULES_DIR}/grpcall
+install(FILES cmake/Modules/grgrpcConfig.cmake
+    DESTINATION ${CMAKE_MODULES_DIR}/grgrpc
 )
