--- gr-grpc/docs/README.grpcall
@@ -1,11 +1,0 @@
-This is the grpcall-write-a-block package meant as a guide to building
-out-of-tree packages. To use the grpcall blocks, the Python namespaces
-is in 'grpcall', which is imported as:
-
-    import grpcall
-
-See the Doxygen documentation for details about the blocks available
-in this package. A quick listing of the details can be found in Python
-after importing by using:
-
-    help(grpcall)
