+++ gr-grpc/docs/doxygen/other/main_page.dox
@@ -1,8 +1,8 @@
 /*! \mainpage
 
-Welcome to the GNU Radio GRPCALL Block
+Welcome to the GNU Radio GRGRPC Block
 
-This is the intro page for the Doxygen manual generated for the GRPCALL
+This is the intro page for the Doxygen manual generated for the GRGRPC
 block (docs/doxygen/other/main_page.dox). Edit it to add more detailed
 documentation about the new GNU Radio modules contained in this
 project.
