+++ gr-grpc/grc/CMakeLists.txt
@@ -17,6 +17,6 @@
 # the Free Software Foundation, Inc., 51 Franklin Street,
 # Boston, MA 02110-1301, USA.
 install(FILES
-    grpcall_msg_sink.xml
+    grgrpc_msg_sink.xml
     DESTINATION share/gnuradio/grc/blocks
 )
